CollageMaker/requirements.txt
numpy~=1.20.3
Pillow~=8.1.0